Honor settings for including / excluding cd9660, ufs, ext2fs and msdos.
The Makefile gives the impression that ext2fs and msdos were excluded
(they weren't) and that you could exclude cd9660 and ufs support (you
couldn't). Allow those to be excluded.

We need to look, in the future, at trimming the number of supported
filesystems, and this will make that easier.


git-svn-id: https://svn.freebsd.org/base/head@329726 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f