MFC 323889: Place the AAD before the plaintext/ciphertext for CIOCRYPTAEAD.
Software crypto implementations don't care how the buffer is laid out,
but hardware implementations may assume that the AAD is always before
the plain/cipher text and that the hash/tag is immediately after the end
of the plain/cipher text.

In particular, this arrangement matches the layout of both IPSec packets
and TLS frames.  Linux's crypto framework also assumes this layout for
AEAD requests.

Sponsored by:	Chelsio Communications


git-svn-id: https://svn.freebsd.org/base/stable/11@329773 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f