lualoader: Remove remnants of testing...
twiddle_pos didn't need to be a module-scope local, since it's going to get
reset with every read anyways- it was left-over from other things.

screen.movecursor with a y=-1 setting was from a test of movecursor,
resulting in the twiddle characters being drawn going up the console and
looking quite funky.


git-svn-id: https://svn.freebsd.org/base/head@330083 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f