MFC r306000
[iwm] Remove unused field from iwm_rx_data. Use uint32_t instead of enum type.

The wantresp field in struct iwm_rx_data has never been used for anything,
so we can just delete it.

Apparently struct iwm_sf_cfg_cmd was compiled correctly (using a 32bit
value to represent the enum), but it still seems like a very bad idea to use
an enum type in a __packed struct.


git-svn-id: https://svn.freebsd.org/base/stable/11@330146 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f