liblua: Use putc instead of printf for printc
printc does not need the features or the overhead of printf. It does not
take formatting strings, and it pipes the single string argument through an
"%s" format.

Instead, use putc directly. This pipes the string through in its entirety as
a series of 'unsigned char's, generally straight to the console emulator.

Discussed with:	tsoome


git-svn-id: https://svn.freebsd.org/base/head@330284 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f