MFC r302519:
Audit the file-descriptor number argument for openat(2).  Remove a comment
about the desirability of auditing the number, as it was in fact in the
wrong place (in the common path for open(2) and openat(2), and only the
latter accepts a file-descriptor argument).  Where other ABIs support
openat(2), it may be necessary to do additional argument auditing as it is
not performed in kern_openat(9).

MFC after:	3 days


git-svn-id: https://svn.freebsd.org/base/stable/11@330343 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f